Implemented improved large gallery with 55 rooms
Finally implemented a tile-based gallery layout
to greatly expand the image capacity.

Also prettyfied the lighting with the help of an
HDR from Greg Zaal @ Poly Haven (Licensed CC0)

Rooms as well as painting slots are generated
asynchronously from the renderer but in order,
always generating first the closest element to (0,0)
in hamiltonian distance. All slots are put on
a global list, which is not an issue because
you can know their chunk by accessing the 'parent'
attr.